services/TenantService/pyproject.toml
[project]
name = "tenantservice"
version = "0.1.0"
description = "Tenant management service for Legal Document Analyzer"
authors = [
    {name = "yithril",email = "yithril@gmail.com"}
]
requires-python = ">=3.13,<4.0"
dependencies = [
    "sqlalchemy (>=2.0.0,<3.0.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1" 